requirements.txt
transformers
torch
tqdm
huggingface_hub
torchvision
numpy>=1.21.0
sentencepiece==0.1.99
Pillow>=10.1.0
google-generativeai
colornamer==0.2.3
scikit_learn>=1.4.0
webcolors==1.13